[ads-collector] move source to src and store outputs in dist
