+++ .gitignore
@@ -6,3 +6,4 @@
 *.xlsx
 *.json
 __pycache__
+dist/
