+++ README.md
@@ -34,15 +34,17 @@ AD_PROVIDERS=meta,google
 STORAGES=csv,excel,bigquery,mysql
 ```
 
-Provide a Google Cloud service account JSON key and update the path in `run.py` if needed. Specify the collection period with the `--start-date` and `--end-date` arguments when running the script. Both default to yesterday.
+Provide a Google Cloud service account JSON key and update the path in
+`src/run.py` if needed. Specify the collection period with the `--start-date`
+and `--end-date` arguments when running the script. Both default to yesterday.
 
 ## Usage
 Install the dependencies and run the script:
 
 ```
 pip install -r requirements.txt
-python migrate.py  # run once to create/update tables
-python run.py --start-date 2023-01-01 --end-date 2023-01-02
+python src/migrate.py  # run once to create/update tables
+python src/run.py --start-date 2023-01-01 --end-date 2023-01-02
 
 Providers listed in `AD_PROVIDERS` map to classes in the `providers` package.
 
@@ -59,7 +61,7 @@ docker compose up -d
 
 This launches MySQL 8 and stores the data in a named volume `mysql_data`.
 Docker Compose reads the connection variables from your `.env` file so the
-container uses the same settings as `run.py`. The service falls back to the
+container uses the same settings as `src/run.py`. The service falls back to the
 following defaults if a variable is not provided:
 
 - database: `${MYSQL_DATABASE:-ads}`
@@ -67,4 +69,4 @@ following defaults if a variable is not provided:
 - password: `${MYSQL_PASSWORD:-ads_password}`
 - root password: `${MYSQL_ROOT_PASSWORD:-root}`
 
-Set matching values in your `.env` file so `run.py` can connect.
+Set matching values in your `.env` file so `src/run.py` can connect.
